Separate Peer and PeerStore, make app.Run non-blocking
