+++ peer/peer.go
@@ -1,7 +1,6 @@
 package peer
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -26,7 +25,7 @@ const (
 	DefaultIP = "127.0.0.1"
 	// MessageWaitTime is the amount of time the dispatcher should wait before
 	// attempting to read from the connection again when no data was received
-	MessageWaitTime = time.Second * 5
+	MessageWaitTime = time.Second * 2
 	// MaxPeers is the maximum number of peers we can be connected to at a time
 	MaxPeers = 50
 	// PeerSearchWaitTime is the amount of time the maintainConnections goroutine
@@ -35,161 +34,14 @@ const (
 	PeerSearchWaitTime = time.Second * 10
 )
 
-// PeerStore is a thread-safe container for all the peers we are currently
-// connected to. It maps remote peer listen addresses to Peer objects.
-type PeerStore struct {
-	peers                 map[string]*Peer
-	ListenAddr            string
-	defaultRequestHandler RequestHandler
-	defaultPushHandler    PushHandler
-	lock                  *sync.RWMutex
-}
+// ResponseHandler is any function that handles a response to a request.
+type ResponseHandler func(*msg.Response)
 
-// NewPeerStore returns an initialized peerstore.
-func NewPeerStore(la string) *PeerStore {
-	return &PeerStore{
-		peers:      make(map[string]*Peer, 0),
-		ListenAddr: la,
-		lock:       &sync.RWMutex{},
-	}
-}
+// PushHandler is any function that handles a push message.
+type PushHandler func(*msg.Push)
 
-// Connect attempts to establish a connection with a peer given its listen
-// address (in the form <IP address>:<TCP port>). If successful returns the
-// peer, otherwise returns error.
-func Connect(address string, ps *PeerStore) (*Peer, error) {
-	c, err := conn.Dial(address)
-	if err != nil {
-		return nil, err
-	}
-	ps.ConnectionHandler(c)
-	p := ps.Get(c.RemoteAddr().String())
-	if p == nil {
-		// This will only be the case if we exchangeListedAddrs fails
-		return nil, errors.New("Failed to exchange listen addresses with peer")
-	}
-	return p, nil
-}
-
-// ConnectionHandler is called when a new connection is opened with us by a
-// remote peer. It will create a dispatcher and message handlers to handle
-// retrieving messages over the new connection and sending them to App.
-func (ps *PeerStore) ConnectionHandler(c net.Conn) {
-	// Before we can continue we must exchange listen addresses
-	addr, err := exchangeListenAddrs(c, PeerSearchWaitTime, ps.ListenAddr)
-	if err != nil {
-		log.WithError(err).Error("Failed to retrieve peer listen address")
-		return
-	}
-	p := New(c, ps, addr)
-
-	// If we are already at MaxPeers, disconnect from a peer to connect to a new
-	// one. This way nobody gets choked out of the network because everybody
-	// happens to be fully connected.
-	if p.Store.Size() >= MaxPeers {
-		p.Store.RemoveRandom()
-	}
-	p.Store.Add(p)
-
-	go p.Dispatch()
-	log.Infof("Connected to %s", p.ListenAddr)
-}
-
-// Add synchronously adds the given peer to the peerstore
-func (ps *PeerStore) Add(p *Peer) {
-	ps.lock.Lock()
-	defer ps.lock.Unlock()
-	ps.peers[p.ListenAddr] = p
-}
-
-// Remove synchronously removes the given peer from the peerstore
-func (ps *PeerStore) Remove(addr string) {
-	ps.lock.Lock()
-	defer ps.lock.Unlock()
-	delete(ps.peers, addr)
-}
-
-// RemoveRandom removes a random peer from the given PeerStore
-func (ps *PeerStore) RemoveRandom() {
-	ps.lock.Lock()
-	defer ps.lock.Unlock()
-	for _, p := range ps.peers {
-		delete(ps.peers, p.ListenAddr)
-		break
-	}
-}
-
-// Get synchronously retreives the peer with the given id from the peerstore
-func (ps *PeerStore) Get(addr string) *Peer {
-	ps.lock.RLock()
-	defer ps.lock.RUnlock()
-	return ps.peers[addr]
-}
-
-// SetDefaultRequestHandler will ensure that all new peers created who's
-// RequestHandlers have not been set will use the given request handler by default
-// until it is overridden by the call to SetRequestHandler().
-func (ps *PeerStore) SetDefaultRequestHandler(rh RequestHandler) {
-	ps.defaultRequestHandler = rh
-}
-
-// SetDefaultPushHandler will ensure that all new peers created who's
-// PushHandlers have not been set will use the given request handler by default
-// until it is overridden by the call to SetPushHandler().
-func (ps *PeerStore) SetDefaultPushHandler(ph PushHandler) {
-	ps.defaultPushHandler = ph
-}
-
-// Broadcast sends the given push message to all peers in the PeerStore at the
-// time this function is called. Note that if we fail to write the push message
-// to a peer the failure is ignored. Generally this is okay, because push
-// messages sent via Broadcast() should be propagated by other peers.
-func (ps *PeerStore) Broadcast(push msg.Push) {
-	ps.lock.RLock()
-	defer ps.lock.RUnlock()
-	for _, p := range ps.peers {
-		p.Push(push)
-	}
-}
-
-// PeerInfoHandler will handle the response to a PeerInfo request by attempting
-// to establish connections with all new peers in the given response Resource.
-func (ps *PeerStore) PeerInfoHandler(res *msg.Response) {
-	peers := res.Resource.([]string)
-	strPeers, _ := json.Marshal(peers)
-	log.Debugf("Found peers %s", string(strPeers))
-	for i := 0; i < len(peers) && ps.Size() < MaxPeers; i++ {
-		if ps.Get(peers[i]) != nil || peers[i] == ps.ListenAddr {
-			// We are already connected to this peer. Skip it.
-			continue
-		}
-
-		p, err := Connect(peers[i], ps)
-		if err != nil {
-			log.WithError(err).Errorf("Failed to dial peer %s", peers[i])
-		}
-		log.Infof("Connected to %s", p.ListenAddr)
-	}
-}
-
-// Addrs returns the list of addresses of the peers in the peerstore in the form
-// <IP addr>:<port>
-func (ps *PeerStore) Addrs() []string {
-	ps.lock.RLock()
-	defer ps.lock.RUnlock()
-	addrs := make([]string, 0)
-	for addr := range ps.peers {
-		addrs = append(addrs, addr)
-	}
-	return addrs
-}
-
-// Size synchornously returns the number of peers in the PeerStore
-func (ps *PeerStore) Size() int {
-	ps.lock.RLock()
-	defer ps.lock.RUnlock()
-	return len(ps.peers)
-}
+// RequestHandler is any function that returns a response to a request.
+type RequestHandler func(*msg.Request) msg.Response
 
 // Peer represents a remote peer we are connected to.
 type Peer struct {
@@ -202,15 +54,6 @@ type Peer struct {
 	lock             sync.RWMutex
 }
 
-// ResponseHandler is any function that handles a response to a request.
-type ResponseHandler func(*msg.Response)
-
-// PushHandler is any function that handles a push message.
-type PushHandler func(*msg.Push)
-
-// RequestHandler is any function that returns a response to a request.
-type RequestHandler func(*msg.Request) msg.Response
-
 // New returns a new Peer
 func New(c net.Conn, ps *PeerStore, listenAddr string) *Peer {
 	return &Peer{
@@ -221,6 +64,24 @@ func New(c net.Conn, ps *PeerStore, listenAddr string) *Peer {
 		pushHandler:      ps.defaultPushHandler,
 		responseHandlers: make(map[string]ResponseHandler),
 	}
+}
+
+// Connect attempts to establish a connection with a peer given its listen
+// address (in the form <IP address>:<TCP port>). If successful returns the
+// peer, otherwise returns error. Once the connection is established the peer
+// will be added to the given PeerStore and returned.
+func Connect(address string, ps *PeerStore) (*Peer, error) {
+	c, err := conn.Dial(address)
+	if err != nil {
+		return nil, err
+	}
+	ps.ConnectionHandler(c)
+	p := ps.Get(c.RemoteAddr().String())
+	if p == nil {
+		// This will only be the case if we exchangeListedAddrs fails
+		return nil, errors.New("Failed to exchange listen addresses with peer")
+	}
+	return p, nil
 }
 
 // SetRequestHandler will add the given request handler to this peer. The
@@ -253,7 +114,7 @@ func (p *Peer) Dispatch() {
 			} else {
 				if strings.Contains(err.Error(), syscall.ECONNRESET.Error()) || errCount == 3 {
 					log.WithError(err).Infof("Disconnecting from peer %s",
-						p.Connection.RemoteAddr().String())
+						p.ListenAddr)
 					p.Store.Remove(p.ListenAddr)
 					p.Connection.Close()
 					return
@@ -269,7 +130,7 @@ func (p *Peer) Dispatch() {
 		case *msg.Request:
 			if p.requestHandler == nil {
 				log.Errorf("Request received but no request handler set for peer %s",
-					p.Connection.RemoteAddr().String())
+					p.ListenAddr)
 			} else {
 				response := p.requestHandler(message.(*msg.Request))
 				response.Write(p.Connection)
@@ -278,15 +139,16 @@ func (p *Peer) Dispatch() {
 			res := message.(*msg.Response)
 			rh := p.getResponseHandler(res.ID)
 			if rh == nil {
-				log.Error("Dispatcher could not find response handler for response")
+				log.Error("Dispatcher could not find response handler for response on peer",
+					p.ListenAddr)
 			} else {
 				rh(res)
 				p.removeResponseHandler(res.ID)
 			}
 		case *msg.Push:
 			if p.pushHandler == nil {
-				log.Errorf("Push message received but no push handler set for peer %s",
-					p.Connection.RemoteAddr().String())
+				log.Error("Dispatcher could not find push handler for push message on peer",
+					p.ListenAddr)
 			} else {
 				p.pushHandler(message.(*msg.Push))
 			}
@@ -330,10 +192,12 @@ func (ps *PeerStore) MaintainConnections() {
 				ResourceType: msg.ResourcePeerInfo,
 			}
 			p := ps.Get(peerAddrs[i])
-			if p != nil && peerAddrs[i] != ps.ListenAddr {
+			ps.lock.RLock()
+			if p != nil {
 				// Need to do this check in case the peer got removed
 				p.Request(peerInfoRequest, ps.PeerInfoHandler)
 			}
+			ps.lock.RUnlock()
 		}
 		// Looks like we hit peer.MaxPeers. Wait for a while before checking how many
 		// peers we are connected to. We don't want to spin.
@@ -359,6 +223,11 @@ func (p *Peer) getResponseHandler(id string) ResponseHandler {
 	return p.responseHandlers[id]
 }
 
+// Performs a handshake over the given connection allowing us to send our
+// listen address to the remote peer and to receive its litsten address.
+// On success returns remote peer listen address, on failure returns error. If
+// the given duration passes and we havn't received a listen address we return
+// an error.
 func exchangeListenAddrs(c net.Conn, d time.Duration, listenAddr string) (string, error) {
 	addrChan := make(chan string)
 	errChan := make(chan error)
@@ -421,7 +290,8 @@ func exchangeListenAddrs(c net.Conn, d time.Duration, listenAddr string) (string
 	case err := <-errChan:
 		return "", err
 	case <-time.After(d):
-		return "", fmt.Errorf("Failed to exchange listen addresses with %s", c.RemoteAddr().String())
+		return "", fmt.Errorf("Failed to exchange listen addresses with %s",
+			c.RemoteAddr().String())
 	}
 }
